// src/app/model/Trainer.ts
import {Material} from './Material';

export class Trainer {
    id: number;
    name: string;
    gender: string;
    placeOb: string;
    birthDate: Date;
    address: string;
    email: string;
    phoneNumber: string;
    username: string;
    password: string;
    material: Material;
    createdAt;
    updateAt;

}